+++ src/actions/shynees.js
@@ -1,7 +1,5 @@
 import {
   getShyneesAroundRequest,
-  getShyneeProfileRequest,
-  getShyneeSettingsRequest
 } from '../request/shynees';
 
 export const SHYNEES_AROUND_REQUEST = 'SHYNEES_AROUND_REQUEST';
@@ -20,46 +18,6 @@ export const getShyneesAround = () => (dispatch) => {
     }))
     .catch(error => dispatch({
       type: SHYNEES_AROUND_FAILURE,
-      payload: error
-    }));
-};
-
-export const SHYNEE_PROFILE_REQUEST = 'SHYNEE_PROFILE_REQUEST';
-export const SHYNEE_PROFILE_SUCCESS = 'SHYNEE_PROFILE_SUCCESS';
-export const SHYNEE_PROFILE_FAILURE = 'SHYNEE_PROFILE_FAILURE';
-
-export const getShyneeProfile = (shyneeId) => (dispatch) => {
-  dispatch({
-    type: SHYNEE_PROFILE_REQUEST
-  });
-
-  return getShyneeProfileRequest(shyneeId)
-    .then((shynees) => dispatch({
-      type: SHYNEE_PROFILE_SUCCESS,
-      payload: shynees
-    }))
-    .catch(error => dispatch({
-      type: SHYNEE_PROFILE_FAILURE,
-      payload: error
-    }));
-};
-
-export const SHYNEE_SETTINGS_REQUEST = 'SHYNEE_SETTINGS_REQUEST';
-export const SHYNEE_SETTINGS_SUCCESS = 'SHYNEE_SETTINGS_SUCCESS';
-export const SHYNEE_SETTINGS_FAILURE = 'SHYNEE_SETTINGS_FAILURE';
-
-export const getShyneeSettings = (shyneeId) => (dispatch) => {
-  dispatch({
-    type: SHYNEE_SETTINGS_REQUEST
-  });
-
-  return getShyneeSettingsRequest(shyneeId)
-    .then((shynees) => dispatch({
-      type: SHYNEE_SETTINGS_SUCCESS,
-      payload: shynees
-    }))
-    .catch(error => dispatch({
-      type: SHYNEE_SETTINGS_FAILURE,
       payload: error
     }));
 };
